backend/requirements.txt
plotly
Flask
Flask-Cors
setuptools
wheel
numpy
cython
nibabel
git+https://github.com/gallantlab/pycortex.git
scikit-image
templateflow
scipy
SQLAlchemy
Flask-SQLAlchemy
Flask-Migrate
psycopg2-binary
python-dotenv